pyproject.toml
[project]
name = "RagDocumentLoader"
version = "0.1.0"
dependencies = [
    "dotenv>=0.9.9",
    "psycopg2-binary>=2.9.11",
    "pyyaml>=6.0.3",
]
